reorder imports and tests
